+++ libraries/LUHN/LUHN.h
@@ -2,7 +2,7 @@
 //
 //    FILE: LUHN.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.2.0
 //    DATE: 2022-12-24
 // PURPOSE: Arduino Library for calculating LUHN checksum.
 //     URL: https://github.com/RobTillaart/LUHN
@@ -10,7 +10,7 @@
 
 #include "Arduino.h"
 
-#define LUHN_LIB_VERSION                      (F("0.1.1"))
+#define LUHN_LIB_VERSION                      (F("0.2.0"))
 
 
 class LUHN
@@ -22,6 +22,7 @@ public:
   //        buffer == \0 terminated
   bool      isValid(const char * buffer);
   bool      isValid(char * buffer);
+  char      generateChecksum(const char * buffer);
   char      generateChecksum(char * buffer);
 
   //  GENERATE A PRODUCT ID WITH LUHN CHECKSUM
@@ -31,15 +32,16 @@ public:
   //  STREAM INTERFACE
   char      add(char c);
   char      reset();
-
+  uint32_t  count();
 
 protected:
   uint32_t  m_w = 1;  //  random generator parameter
   uint32_t  m_z = 2;  //  random generator parameter
   uint8_t   Marsaglia_mod10();
-  
-  uint8_t   _luhn = 0;
-  uint16_t  _count = 0;
+
+  uint16_t  _luhnEven = 0;
+  uint16_t  _luhnOdd = 0;
+  uint32_t  _count = 0;
 };